services/rag_api/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
pydantic-settings>=2.8.0
langchain>=0.3.0
langchain-openai>=0.3.0
langgraph>=0.2.0
qdrant-client>=1.14.0
python-multipart>=0.0.20
httpx>=0.28.0
python-dotenv>=1.0.0
ruff>=0.9.0
black>=24.12.0
isort>=5.13.0

# Phase 2 dependencies - Hybrid Retrieval
sentence-transformers>=3.0.0
transformers>=4.36.0
torch>=2.0.0
numpy>=1.24.0
openai>=1.0.0

# Optional GPU support
# torch-audio  # Uncomment if using audio models
# faiss-cpu>=1.7.0  # Uncomment for efficient similarity search
# faiss-gpu>=1.7.0  # Uncomment for GPU-accelerated search

# Phase 8 dependencies - Evaluation & Metrics
ragas>=0.1.0
datasets>=2.14.0
pandas>=2.0.0
scikit-learn>=1.3.0

# Phase 9 dependencies - Performance & Cost Tuning
pyyaml>=6.0.0
nltk>=3.8.0
aiofiles>=24.0.0

# Phase 10 Dependencies - Logging & Tracing
structlog>=23.1.0
langfuse>=2.0.0
phoenix-evals>=0.8.0
opentelemetry-api>=1.20.0
opentelemetry-sdk>=1.20.0
opentelemetry-instrumentation-fastapi>=0.41b0
opentelemetry-exporter-otlp>=1.20.0
prometheus-client>=0.19.0
psutil>=5.9.0

# Development and testing
pytest>=7.0.0
pytest-asyncio>=0.21.0
